Assign columns to variables and move obtainLabels function to utils
